This is the upload of the files for the Food Truck
Here are the files for the Food Truck app that has the PHP code into it.  This includes the menu items which are classes and objects and the food truck app that implements the menu into it and calculates the total cost of all items purchased.  [A style sheet is also included for design purposes of the app]